+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import registerServiceWorker from './registerServiceWorker';
+import './stylesheets/index.css';
+import App from './components/App';
+import registerServiceWorker from './js/registerServiceWorker';
 
 import 'bootstrap/dist/css/bootstrap.css';
 import $ from 'jquery';
